+++ src/Environments/WorldEnvironment.js
@@ -93,6 +93,10 @@ class WorldEnvironment extends Environment{
             this.largest_cell_count = organism.anatomy.cells.length;
     }
 
+    canAddOrganism() {
+        return this.organisms.length < Hyperparams.maxOrganisms || Hyperparams.maxOrganisms < 0;
+    }
+
     averageMutability() {
         if (this.organisms.length < 1)
             return 0;
@@ -193,8 +197,12 @@ class WorldEnvironment extends Environment{
     loadRaw(env) { // species name->stats map, evolution controls, 
         this.organisms = [];
         FossilRecord.clear_record();
-        this.resizeGridColRow(this.grid_map.cell_size, env.grid.cols, env.grid.rows)
+        let cell_size = env.grid.cell_size ? env.grid.cell_size : this.grid_map.cell_size;
+        this.resizeGridColRow(cell_size, env.grid.cols, env.grid.rows)
         this.grid_map.loadRaw(env.grid);
+        for (let wall of env.grid.walls) {
+            this.walls.push(this.grid_map.cellAt(wall.c, wall.r));
+        }
 
         // create species map
         let species = {};
@@ -218,6 +226,7 @@ class WorldEnvironment extends Environment{
                 s.anatomy = org.anatomy;
                 s.calcAnatomyDetails();
             }
+            s.name = orgRaw.species_name;
             org.species = s;
         }
         for (let name in species)